+++ versions.tf
@@ -1,7 +1,10 @@
 terraform {
-  required_version = ">= 0.12"
+  required_version = ">= 0.13"
 
   required_providers {
-    azurerm = ">= 1.31"
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = ">= 1.31"
+    }
   }
 }